Fix a cast to an unsigned int. Fix the test suite because sm now returns real headers (without the empty line at the end):  it is the sed "Q" option which should be used, and not "q".

+++ nnrpd/group.c
@@ -137,7 +137,7 @@ CMDgroup(int ac, char *av[])
 			ARThigh = prev;
                         /* No need to update the count since it is only
                          * an estimate but make sure it is not too high. */
-                        if (count > ARThigh - ARTlow)
+                        if ((unsigned int)count > ARThigh - ARTlow)
                             count = ARThigh - ARTlow + 1;
 			break;
 		    }
